Move TaylorSeries to new folder
